不要ファイルの全削除とソース整理を行った。フレームワークの原型は完成？

+++ DX22Base/_model.cpp
@@ -68,10 +68,10 @@ void Model::MakeMesh(const void* ptr, float scale, Flip flip)
 		MeshBuffer::Description desc = {};
 		desc.pVtx =		m_meshes[i].vertices.data();
 		desc.vtxSize =	sizeof(Vertex);
-		desc.vtxCount =	m_meshes[i].vertices.size();
+		desc.vtxCount =	(UINT)m_meshes[i].vertices.size();
 		desc.pIdx =		m_meshes[i].indices.data();
 		desc.idxSize =	sizeof(unsigned long);
-		desc.idxCount =	m_meshes[i].indices.size();
+		desc.idxCount =	(UINT)m_meshes[i].indices.size();
 		desc.topology =	D3D11_PRIMITIVE_TOPOLOGY_TRIANGLELIST;
 		m_meshes[i].pMesh = new MeshBuffer(desc);
